tests/test_migration.py
import pytest
from unittest.mock import patch, MagicMock
from etl.schema.migration import apply_schema_migration

@patch('migration.psycopg2.connect')
def test_apply_schema_migration(mock_psycopg2_connect):
    mock_conn = MagicMock()
    mock_cursor = MagicMock()
    mock_psycopg2_connect.return_value = mock_conn
    mock_conn.cursor.return_value = mock_cursor

    conn_params = {
        'dbname': 'stock_warehouse',
        'user': 'postgres',
        'password': 'ies123@IES123',
        'host': 'localhost',
        'port': '5432'
}

    try:
        apply_schema_migration(conn_params)
        mock_psycopg2_connect.assert_called_once_with(**conn_params)
        mock_cursor.execute.assert_called()  # Ensure SQL execution calls
        assert True  # Placeholder assertion
    except Exception as e:
        pytest.fail(f"Exception raised: {e}")
